Move headers to include file
