+++ CMakeLists.txt
@@ -109,5 +109,6 @@ if(MM_USE_EXTRA_WARNINGS)
     endif()
 endif()
 
+add_subdirectory(include)
 add_subdirectory(src)
 add_subdirectory(test)
