+++ src/CMakeLists.txt
@@ -1,4 +1,3 @@
-add_subdirectory(common)
 add_subdirectory(sequential)
 add_subdirectory(parallel)
 add_subdirectory(traces)
