+++ src/traces/include/traces/traces.hpp
@@ -2,9 +2,9 @@
 
 #include <vector>
 
-#include "../common/logger.hpp"
-#include "../common/status.hpp"
-#include "../common/types.hpp"
+#include "common/logger.hpp"
+#include "common/status.hpp"
+#include "common/types.hpp"
 
 enum class TraceOperator {
     insert,
